pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name= "ClimateGraph"
version = "0.1.0"
description = "A CLI tool for graphing Climate Data"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{ name = "Antonio Andrés Torga Mellado", email = "antonio.torga@ug.uchile.cl" }]
dependencies = [
        "typer>=0.12",
        "xarray",
        "netcdf4",
        "numpy",
        "matplotlib",
        "pyyaml",
        "pyresample",
        "cartopy",
        "pydantic",
]

[project.optional-dependencies]
dev = [
        "black"
]

test= [
        "pytest"
]
[project.scripts]
ClimateGraph= "ClimateGraph.cli:app"